refactor dcap_retrieve_pckid to lib & main
so that this code could be reuse elsewhere
